feat(api): create models for all database tables
Models are structs that represent database entities (tables) and
encapsulate associated logic, which we'll add gradually.

Maybe we want to interact with the database through these models, like
with an ORM?

It's a little much, but we need quite a few of them to satisfy the
foreign key constraints, so at that point I might as well add all of
them.