fix(api/routes): remove superflous write to response headers
We were unsure whether it is necessary to manually add the `status: ok`
header on each response. Judging from this warning message I'm seeing
when running the app, I'm guessing it isn't.

```
2024/12/16 10:39:28 http: superfluous response.WriteHeader call from dreamchasers/routes.ExecuteRouteHandler (routes.go:41)
```

Some quick googling also told me that the success header is added
automatically by default, and also that managing response status codes
and other headers might be kind of complicated. I'm sure we'll have to
get back to this eventually.