feat(frontend): get map POIs from API instead of hard coded list
POIs ("plekken") are now fetched from the API, falling back on the hard
coded list if the request failed. One problem is that the API's URL
being fetched depends on the hosting solution. Keeping it at
`http://localhost:4000` is not an option. Perhaps if we finish the
docker setup, we could use the configured hostname/port of the container
instead. That way it should work when developing locally using the
docker-compose configuration and when deployed to the server.

Also, facilities such as wash- and bathrooms really aren't
accommodations at all. I had not accounted for those in the database
(nor has anyone pointed out the flaw to me), so the backend will have to
change to better handle those. For now I've added them to accommodations
table with a price of 0, but it really should be fixed.